+++ app/src/test/java/ru/perm/v/infomaximum/parser/jsonreader/StatCreatorByProductsTest.java
@@ -15,27 +15,27 @@ public class StatCreatorByProductsTest {
     @Test
     public void testGetMaxWeight() {
         Product product1 = new Product();
-        product1.setWeight(111);
+        product1.setWeight(111L);
         Product product2 = new Product();
-        product2.setWeight(222);
+        product2.setWeight(222L);
         List<Product> productList = Arrays.asList(product1, product2);
         StatCreatorByListProducts statCreatorByProducts = new StatCreatorByListProducts();
 
-        Integer maxWeight = statCreatorByProducts.getMaxWeight(productList);
+        Long maxWeight = statCreatorByProducts.getMaxWeight(productList);
 
-        assertEquals(Integer.valueOf(222), maxWeight);
+        assertEquals(Long.valueOf(222), maxWeight);
     }
 
     @Test
     public void testGetMinWeight() {
         Product product1 = new Product();
-        product1.setWeight(111);
+        product1.setWeight(111L);
         Product product2 = new Product();
-        product2.setWeight(222);
+        product2.setWeight(222L);
         List<Product> productList = Arrays.asList(product1, product2);
         StatCreatorByListProducts statCreatorByProducts = new StatCreatorByListProducts();
 
-        Integer minWeight = statCreatorByProducts.getMinWeight(productList);
+        Long minWeight = statCreatorByProducts.getMinWeight(productList);
 
         assertEquals(Integer.valueOf(111), minWeight);
     }
@@ -45,23 +45,23 @@ public class StatCreatorByProductsTest {
         String GRP1 = "GRP1";
         Product product11 = new Product();
         product11.setGrp(GRP1);
-        product11.setWeight(111);
+        product11.setWeight(111L);
         Product product12 = new Product();
         product12.setGrp(GRP1);
-        product12.setWeight(112);
+        product12.setWeight(112L);
 
         String GRP2 = "GRP2";
         Product product21 = new Product();
         product21.setGrp(GRP2);
-        product21.setWeight(222);
+        product21.setWeight(222L);
         Product product22 = new Product();
         product22.setGrp(GRP2);
-        product22.setWeight(333);
+        product22.setWeight(333L);
 
         List<Product> productList = Arrays.asList(product11, product12, product21, product22);
         StatCreatorByListProducts statCreatorByProducts = new StatCreatorByListProducts();
 
-        Map<String, Integer> weightByGroup = statCreatorByProducts.getSumWegthByGroup(productList);
+        Map<String, Long> weightByGroup = statCreatorByProducts.getSumWegthByGroup(productList);
 
         assertEquals(Integer.valueOf(223), weightByGroup.get(GRP1));
         assertEquals(Integer.valueOf(555), weightByGroup.get(GRP2));
@@ -73,29 +73,29 @@ public class StatCreatorByProductsTest {
         String TYPE1 = "TYPE1";
 
         Product productG1T1N1 = new Product();
-        productG1T1N1.setNum(1);
+        productG1T1N1.setNum(1L);
         productG1T1N1.setGrp(GRP1);
         productG1T1N1.setType(TYPE1);
 
         Product productG1T1N2 = new Product();
-        productG1T1N2.setNum(2);
+        productG1T1N2.setNum(2L);
         productG1T1N2.setGrp(GRP1);
         productG1T1N2.setType(TYPE1);
 
         String GRP2 = "GRP2";
         String TYPE2 = "TYPE2";
         Product productG2T2N3 = new Product();
-        productG2T2N3.setNum(3);
+        productG2T2N3.setNum(3L);
         productG2T2N3.setGrp(GRP2);
         productG2T2N3.setType(TYPE2);
 
         Product productG2T2N4 = new Product();
-        productG2T2N4.setNum(4);
+        productG2T2N4.setNum(4L);
         productG2T2N4.setGrp(GRP2);
         productG2T2N4.setType(TYPE2);
 
         Product otherProduct = new Product();
-        otherProduct.setNum(5);
+        otherProduct.setNum(5L);
         otherProduct.setGrp("-");
         otherProduct.setType("-");
